requirements.txt
<<<<<<< HEAD
discord.py
python-dotenv
=======
discord.py
python-dotenv
>>>>>>> 8267c46fc0b19acd67be4db332cbd2f23e749d22
